Kan no flytte denne implementasjonen ut av prodkoden og over i testkoden
