+++ pensjon-brevbaker/build.gradle.kts
@@ -126,5 +126,6 @@ dependencies {
     testImplementation(libs.pdfbox)
 
     testImplementation(testFixtures(project(":brevbaker")))
+    testImplementation(testFixtures(project(":brevbaker-dsl")))
 }
