+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/brev/template/render/PensjonLatexITest.kt
@@ -6,6 +6,7 @@ import kotlinx.coroutines.runBlocking
 import no.nav.brev.brevbaker.Fixtures
 import no.nav.brev.brevbaker.PDF_BUILDER_URL
 import no.nav.brev.brevbaker.TestTags
+import no.nav.brev.brevbaker.createTemplate
 import no.nav.brev.brevbaker.renderTestPDF
 import no.nav.pensjon.brev.latex.LaTeXCompilerService
 import no.nav.pensjon.brev.template.*
